requirements.txt
APScheduler==3.6.3
async-generator==1.10
async-lru==1.0.1
cycler==0.10.0
gitdb==4.0.2
GitPython==3.1.0
i18ndude==5.3.2
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.3
pyaes==1.6.1
pyparsing==2.4.7
Pyrogram===0.16.0.asyncio
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2019.3
schedule==0.6.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
smmap==3.0.1
TgCrypto==1.2.0
tzlocal==2.0.0
zope.i18nmessageid==5.0.1
zope.interface==5.0.2
zope.tal==4.4
